rename scan as registry
