+++ internal/command/registry/sbom.go
@@ -1,4 +1,4 @@
-package scan
+package registry
 
 import (
 	"context"
