+++ internal/command/registry/scantron.go
@@ -1,4 +1,4 @@
-package scan
+package registry
 
 import (
 	"context"
